Fixed an issue where the app crashed after entering a decimal value as the quantity to be added to the receiving and reverted a theme color I had placed

+++ app/src/main/java/com/scannerapp/cdihandheldscannerviewactivity/ReceivingProductsIntoBin/ReceivingProductsDetailsFragment.kt
@@ -185,7 +185,7 @@ class ReceivingProductsDetailsFragment : Fragment() {
                     if (newExpirationDate != null && DateUtils.isValidDate(expirationDate)) {
                         val expirationDate = newExpirationDateEditText.text.toString()
                         val newLotNumber = newLotEditText.text.toString()
-                        val quantityToAddToDoor = quantityEditText.text.toString().toInt()
+                        val quantityToAddToDoor = quantityEditText.text.toString().toFloat().toInt()
                         val itemNumber = itemNumberTextView.text.toString()
                         val itemName = itemNameTextView.text.toString()
                         val doorBin = viewModel.currentlyChosenDoorBin.value!!.bin_number
